Added support for running unit tests via Ant, and for including extra 3rd party libraries on the robot (tho we currently have none of these) and in simulation. To make this easier, the testing libraries were moved to `libs/test` and any 3rd party libraries required by our robot code will go in `libs/robot`.
The build file was changed in such a way to make it very easy for WPI to incorporate our changes directly into the standard WPILib build file. The only thing teams would then need is a few custom properties (perhaps even optional) properties in the local `build.properties` file that specify the directories for the test and extra robot libraries (which can be empty or non-existant, btw).

Also created several shared launch configurations to make it easier for all developers to use the same recipes.